fix warning message test secrets on all platforms
