+++ dlt/destinations/impl/filesystem/sql_client.py
@@ -24,6 +24,8 @@ from dlt.common.configuration.specs import (
 )
 from dlt.destinations.utils import is_compression_disabled
 
+from pathlib import Path
+
 SUPPORTED_PROTOCOLS = ["gs", "gcs", "s3", "file", "memory", "az", "abfss"]
 
 if TYPE_CHECKING:
@@ -80,17 +82,22 @@ class FilesystemSqlClient(DuckDbSqlClient):
         if persistent and self.memory_db:
             raise Exception("Creating persistent secrets for in memory db is not allowed.")
 
-        current_secret_directory = self._conn.sql(
-            "SELECT current_setting('secret_directory') AS secret_directory;"
-        ).fetchone()[0]
+        secrets_path = Path(
+            self._conn.sql(
+                "SELECT current_setting('secret_directory') AS secret_directory;"
+            ).fetchone()[0]
+        )
 
-        # TODO: what is with windows?
-        is_default_secrets_directory = current_secret_directory.endswith("/.duckdb/stored_secrets")
+        is_default_secrets_directory = (
+            len(secrets_path.parts) >= 2
+            and secrets_path.parts[-1] == "stored_secrets"
+            and secrets_path.parts[-2] == ".duckdb"
+        )
 
         if is_default_secrets_directory and persistent:
             logger.warn(
                 "You are persisting duckdb secrets but are storing them in the default folder"
-                f" {current_secret_directory}. These secrets are saved there unencrypted, we"
+                f" {secrets_path}. These secrets are saved there unencrypted, we"
                 " recommend using a custom secret directory."
             )
