+++ pytest.ini
@@ -1,7 +1,7 @@
 [pytest]
 pythonpath= dlt docs/website/docs
 norecursedirs= .direnv .eggs build dist
-addopts= -v --showlocals --durations 10
+addopts= -v --showlocals --durations 10 --capture=tee-sys
 xfail_strict= true
 log_cli= 1
 log_cli_level= INFO
